26_mlops/05day/total_mlops/api_server/pyproject.toml
[tool.poetry]
name = "ml_app"
version = "0.0.1"
description = ""
authors = ["hyeon <dev.dc.hyeon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
fastapi = {extras = ["all"], version = "^0.111.1"}
kiwipiepy = "^0.18.0"
scikit-learn = "^1.5.1"
mlflow = "^2.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
